feat(textlint-rule-proofdict): support `dictGlob` options
