+++ packages/@proofdict/textlint-rule-proofdict/package.json
@@ -41,7 +41,9 @@
     "localstorage-ponyfill": "^1.0.2",
     "textlint-rule-helper": "^2.1.1",
     "url-join": "^4.0.1",
-    "@textlint/types": "^1.2.2"
+    "@textlint/types": "^1.2.2",
+    "globby": "^10.0.1",
+    "js-yaml": "^3.13.1"
   },
   "devDependencies": {
     "@types/mocha": "^5.2.7",
@@ -53,7 +55,8 @@
     "textlint-scripts": "^3.0.0",
     "ts-node": "^8.4.1",
     "ts-node-test-register": "^8.0.1",
-    "typescript": "^3.6.4"
+    "typescript": "^3.6.4",
+    "@types/js-yaml": "^3.12.1"
   },
   "publishConfig": {
     "access": "public"
