+++ packages/@proofdict/textlint-rule-proofdict/src/create-tester.ts
@@ -1,9 +1,6 @@
 // MIT © 2017 azu
 "use strict";
-import { MODE } from "./mode";
-import { storage } from "./dictionary-storage";
 import { Proofdict, ProofdictTester } from "@proofdict/tester";
-import { RuleOption } from "./RuleOptions";
 
 let currentTester: null | ProofdictTester = null;
 let checkedLastTime = -1;
@@ -46,31 +43,4 @@ export const createTester = ({
         return currentTester;
     }
     return currentTester;
-};
-
-/**
- * @param options
- * @param {string} mode
- * @returns {*}
- */
-export const getDictionary = (options: RuleOption, mode: MODE): any => {
-    // prefer `dictionary` option
-    if (options.proofdict !== undefined) {
-        return options.proofdict;
-    }
-    let proofDictData;
-    // NETWORK
-    if (mode === MODE.NETWORK) {
-        try {
-            const cachedProofdict = storage.getItem("proofdict");
-            proofDictData = JSON.parse(cachedProofdict);
-        } catch (error) {
-            storage.removeItem("proofdict");
-        }
-    }
-    // LOCAL
-    if (mode === MODE.LOCAL) {
-        // TODO: not implemented
-    }
-    return proofDictData;
 };
